Move modifyText to bot.js
